refactor: extract `AESCipher` to `crypto.py`
